Split beacon_server into a module per class
This should make it less of a headache to browse and maintain.
