--- sphinx-doc/infrastructure/beacon_server.rst
@@ -1,5 +1,0 @@
-
-.. automodule:: infrastructure.beacon_server
-   :special-members: __init__
-   :members:
-
