+++ sphinx-doc/infrastructure/beacon_server/base.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.beacon_server.base
+   :special-members: __init__
+   :members:
+
