+++ sphinx-doc/infrastructure/beacon_server/core.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.beacon_server.core
+   :special-members: __init__
+   :members:
+
