+++ sphinx-doc/infrastructure/beacon_server/if_state.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.beacon_server.if_state
+   :special-members: __init__
+   :members:
+
