+++ sphinx-doc/infrastructure/beacon_server/index.rst
@@ -1,0 +1,10 @@
+Beacon Server
+=============
+
+Contents:
+
+.. toctree::
+   :maxdepth: 2
+   :glob:
+
+   *
