+++ sphinx-doc/infrastructure/beacon_server/local.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.beacon_server.local
+   :special-members: __init__
+   :members:
+
