+++ sphinx-doc/infrastructure/beacon_server/rev_obj.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.beacon_server.rev_obj
+   :special-members: __init__
+   :members:
+
