+++ sphinx-doc/simulator/index.rst
@@ -7,8 +7,6 @@ Contents:
    :maxdepth: 3
    :glob:
 
-   *
    application/index
    endhost/index
-   infrastructure/index
    lib/index
