--- sphinx-doc/simulator/infrastructure/index.rst
@@ -1,10 +1,0 @@
-Simulator Infrastructure
-========================
-
-Contents:
-
-.. toctree::
-   :maxdepth: 3
-   :glob:
-
-   *
