--- sphinx-doc/simulator/infrastructure/path_server_sim.rst
@@ -1,4 +1,0 @@
-.. automodule:: simulator.infrastructure.path_server_sim
-   :special-members: __init__
-   :members:
-   :member-order: bysource
