--- sphinx-doc/simulator/infrastructure/router_sim.rst
@@ -1,4 +1,0 @@
-.. automodule:: simulator.infrastructure.router_sim
-   :special-members: __init__
-   :members:
-   :member-order: bysource
