--- sphinx-doc/simulator/simulator.rst
@@ -1,4 +1,0 @@
-.. automodule:: simulator.simulator
-   :special-members: __init__
-   :members:
-   :member-order: bysource
